refactor the python package and separate encoding from opcode
