--- src/encoding/__init__.py
@@ -1,5 +1,0 @@
-from .comparator import *
-from .addition import *
-from .commitment import *
-from .constants import *
-from .utils import *
