+++ src/zkevm_specs/__init__.py
@@ -1,0 +1,2 @@
+from . import encoding
+from . import opcode
