+++ src/zkevm_specs/encoding/lookup.py
