+++ src/zkevm_specs/encoding/typing.py
