+++ src/zkevm_specs/encoding/utils.py
