+++ src/zkevm_specs/opcode/__init__.py
@@ -1,0 +1,2 @@
+from .add import *
+from .comparator import *
